moves print_country method to a new file print_country.rb
this was done in order to reduce UserInterface class length
