+++ lib/user_interface.rb
@@ -10,14 +10,16 @@ require 'byebug'
 require 'tty-spinner'
 require_relative '../lib/scraper_logic'
 require_relative './variables_gen'
+require_relative './print_country'
 
 class UserInterface
-  attr_accessor :scraper_logic, :cases
+  attr_accessor :scraper_logic, :cases, :print_c
 
   def initialize
     @spinner = TTY::Spinner.new
     self.scraper_logic = Scraper.new
     self.cases = VariableGenerator.new
+    self.print_c = PrintCountry.new
     @spinner.auto_spin
     print TTY::Link.link_to('fetching data from: ', 'https://www.worldometers.info/coronavirus/'.blue)
     @doc = Nokogiri::HTML(URI.open('https://www.worldometers.info/coronavirus/'))
@@ -106,73 +108,6 @@ class UserInterface
     )
   end
 
-  def print_second_table_c(country)
-    puts Terminal::Table.new(
-      rows: [
-        [
-          country[:total_recovered],
-          country[:new_recovered],
-          country[:active_cases],
-          country[:serious_critical],
-          country[:total_cases_per_1_milion]
-        ]
-      ],
-      headings: [
-        'Total reco'.bold.green,
-        'New reco'.bold.green,
-        'Active cases'.red,
-        'Serious/critical'.red,
-        'Total cases/1M'.bold.blue
-      ],
-      style: { border_i: '+' }
-    )
-  end
-
-  def print_third_table_c(country)
-    puts Terminal::Table.new(
-      rows: [
-        [
-          country[:deaths_per_1_milion],
-          country[:total_tests],
-          country[:tests_per_1_milion],
-          country[:population],
-          country[:continent]
-        ]
-      ],
-      headings: [
-        'Total deaths/1M'.red,
-        'Total tests'.bold.blue,
-        'Total tests/1M'.bold.blue,
-        'Population'.bold.green,
-        'continent'.bold.green
-      ],
-      style: { border_i: '+' }
-    )
-  end
-
-  def print_country(country)
-    puts Terminal::Table.new(
-      rows: [
-        [
-          country[:name],
-          country[:total_cases],
-          country[:new_cases],
-          country[:total_deaths],
-          country[:new_deaths]
-        ]
-      ], headings: [
-        'Country'.bold.blue,
-        'Total'.red,
-        'New cases'.bold.blue,
-        'Total deaths'.bold.red,
-        'New deaths'.red
-      ],
-      style: { border_i: '+' }
-    )
-    print_second_table_c(country)
-    print_third_table_c(country)
-  end
-
   def ask_user(countries_stats)
     prompt = TTY::Prompt.new
     user_input = prompt.select("Choose 'All' to print 220 indivdual
@@ -181,11 +116,10 @@ class UserInterface
 
     case user_input
     when 'All'
-      print_all(countries_stats)
+      scraper_logic.print_all(countries_stats)
     else
-      print_country(filter_country(countries_stats, user_input))
+      print_c.print_country(scraper_logic.filter_country(countries_stats, user_input))
     end
-    
   end
 
   # print general table method
@@ -195,7 +129,7 @@ class UserInterface
     print_general_table
     print_active_cases
     print_closed_cases
-    scraper_logic.ask_user(@countries_stats)
+    ask_user(@countries_stats)
     font = TTY::Font.new(:doom)
     git_repo = TTY::Link.link_to('if you liked it give a ⭐ t', 'https://github.com/helciodev/covid-19-scraper'.blue)
 
@@ -203,4 +137,3 @@ class UserInterface
     puts git_repo
   end
 end
-
